requirements.txt
kafka-python==2.0.2
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
flask==3.0.0
flask-cors==4.0.0
